Investigate delegatecall (#19)
* ++ delegate-call example

* formatted DeletageCall.sol

* +- address(this) comment explanation for delegatecalls

* ++ DelegateCallHts

* ++ functional non-malicious delegate-hts-call test

* +- hts precompile contracts

* working demo of malicious approval and spend

* demo no issues with delegatecall on ERC20 calls

* +- hedera local node

* ++ comment in RouterHts#depositCallback

+++ tools/hardhat-example/contracts/_precompile-safe-hts/SafeHTS.sol
@@ -541,7 +541,7 @@ library SafeHTS {
 
     modifier nonEmptyExpiry(IHederaTokenService.HederaToken memory token) {
         if (token.expiry.second == 0 && token.expiry.autoRenewPeriod == 0) {
-            token.expiry.autoRenewPeriod = defaultAutoRenewPeriod;
+            token.expiry.autoRenewPeriod = int64(uint64(defaultAutoRenewPeriod));
         }
         _;
     }
